requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
black==24.2.0
blinker==1.7.0
    # via flask
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   flask
distro==1.9.0
    # via openai
exceptiongroup==1.2.0
    # via
    #   anyio
    #   pytest
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
flask==3.0.2
fsspec==2024.2.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via httpcore
html2text==2020.1.16
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via openai
huggingface-hub==0.20.3
    # via
    #   tokenizers
    #   transformers
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
isort==5.13.2
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via
    #   flask
    #   torch
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
    # via jsonschema
lark==1.1.9
lazy-object-proxy==1.10.0
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
mpmath==1.3.0
    # via sympy
mypy-extensions==1.0.0
    # via black
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via transformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.19.3
    # via torch
nvidia-nvjitlink-cu12==12.3.101
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
openai==1.12.0
packaging==23.2
    # via
    #   black
    #   huggingface-hub
    #   pytest
    #   transformers
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via black
pluggy==1.4.0
    # via pytest
pydantic==2.6.1
    # via openai
pydantic-core==2.16.2
    # via pydantic
pytest==8.0.1
    # via pytest-dotenv
pytest-dotenv==0.5.2
python-dotenv==1.0.1
    # via pytest-dotenv
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   transformers
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.4.2
    # via transformers
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sympy==1.12
    # via torch
tokenizers==0.15.2
    # via transformers
tomli==2.0.1
    # via
    #   black
    #   pytest
torch==2.2.0
tqdm==4.66.2
    # via
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.38.1
triton==2.2.0
    # via torch
typing-extensions==4.9.0
    # via
    #   anyio
    #   black
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.2.1
    # via requests
werkzeug==3.0.1
    # via flask
